requirements.txt
speechrecognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
notion-client==2.2.1
aiohttp>=3.8.0
tensorflow>=2.15.0
tensorflow-hub>=0.16.0
librosa>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.4
scipy>=1.11.0
plotly>=5.17.0
hdbscan>=0.8.33